requirements.txt
Django==4.2.14
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-jazzmin==3.0.0
environs==11.0.0
gunicorn==22.0.0
pillow==10.4.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
django-modeltranslation==0.19.5
